requirements.txt
spacy==3.2.1
requests==2.26.0
beautifulsoup4==4.10.0
python-telegram-bot==13.7
